@@ -1,3 +1,3 @@
 import React from 'react';
-export default class UIComponent extends React.PureComponent {
+export default class UIComponent<P = {}, S = {}, SS = any> extends React.PureComponent<P, S, SS> {
 }
