@@ -43,5 +43,5 @@ declare class Text extends PureComponent<PropsTypes> {
     render(): JSX.Element;
 }
 export { Text };
-declare const _default: React.ComponentClass<TextPropTypes, any> | React.FunctionComponent<TextPropTypes>;
+declare const _default: React.ComponentClass<TextPropTypes, any>;
 export default _default;
