@@ -22,6 +22,14 @@ export interface ViewPropTypes extends ViewProps, ContainerModifiers {
      * TODO: probobly isn't needed
      */
     height?: string | number;
+    /**
+     * Experimental: Pass time in ms to delay render
+     */
+    renderDelay?: number;
+    /**
+     * Set background color
+     */
+    backgroundColor?: string;
 }
-declare const _default: React.ComponentClass<ViewPropTypes, any> | React.FunctionComponent<ViewPropTypes>;
+declare const _default: React.ComponentClass<ViewPropTypes, any>;
 export default _default;
