@@ -11,7 +11,12 @@ export default class CustomKeyboardView extends CustomKeyboardViewBase {
     inputRef: PropTypes.object,
     initialProps: PropTypes.object,
     component: PropTypes.string,
-    onItemSelected: PropTypes.func
+    onItemSelected: PropTypes.func,
+    useSafeArea: PropTypes.bool
+  };
+
+  static defaultProps = {
+    useSafeArea: true
   };
 
   constructor(props) {
@@ -42,14 +47,15 @@ export default class CustomKeyboardView extends CustomKeyboardViewBase {
   }
 
   async UNSAFE_componentWillReceiveProps(nextProps) {
-    const {inputRef: nextInputRef, component: nextComponent, initialProps: nextInitialProps} = nextProps;
+    const {inputRef: nextInputRef, component: nextComponent, initialProps: nextInitialProps, useSafeArea} = nextProps;
     const {component} = this.props;
 
     if (nextInputRef && nextComponent !== component) {
       if (nextComponent) {
         TextInputKeyboardManager.setInputComponent(nextInputRef, {
           component: nextComponent,
-          initialProps: nextInitialProps
+          initialProps: nextInitialProps,
+          useSafeArea
         });
       } else {
         TextInputKeyboardManager.removeInputComponent(nextInputRef);
