@@ -110,7 +110,14 @@ class KeyboardAccessoryView extends Component {
      *
      * default: false
      */
-    allowHitsOutsideBounds: PropTypes.bool
+    allowHitsOutsideBounds: PropTypes.bool,
+
+    /**
+     * iOS only.
+     * Whether or not to handle SafeArea
+     * default: true
+     */
+    useSafeArea: PropTypes.bool
   };
 
   static iosScrollBehaviors = IOS_SCROLL_BEHAVIORS;
@@ -236,7 +243,8 @@ class KeyboardAccessoryView extends Component {
       kbInputRef,
       kbComponent,
       onItemSelected,
-      onRequestShowKeyboard
+      onRequestShowKeyboard,
+      useSafeArea
     } = this.props;
 
     return (
@@ -258,6 +266,7 @@ class KeyboardAccessoryView extends Component {
           initialProps={this.processInitialProps()}
           onItemSelected={onItemSelected}
           onRequestShowKeyboard={onRequestShowKeyboard}
+          useSafeArea={useSafeArea}
         />
       </KeyboardTrackingView>
     );
