@@ -3,13 +3,13 @@ import ReactNative, {NativeModules, LayoutAnimation} from 'react-native';
 const CustomInputControllerTemp = NativeModules.CustomInputControllerTemp;
 
 export default class TextInputKeyboardManager {
-  static setInputComponent = (textInputRef, {component, initialProps}) => {
+  static setInputComponent = (textInputRef, {component, initialProps, useSafeArea}) => {
     if (!textInputRef || !CustomInputControllerTemp) {
       return;
     }
     const reactTag = findNodeHandle(textInputRef);
     if (reactTag) {
-      CustomInputControllerTemp.presentCustomInputComponent(reactTag, {component, initialProps});
+      CustomInputControllerTemp.presentCustomInputComponent(reactTag, {component, initialProps, useSafeArea});
     }
   };
 
