@@ -1,49 +1,10 @@
 import baseComponent from '../baseComponent';
-import {Colors, Typography} from '../../style';
+import {Colors} from '../../style';
 
 const BaseComponent = baseComponent(false);
 
 describe('BaseComponent', () => {
-  describe('extractTypographyValue', () => {
-    it('should extract typography value according to typography modifier', () => {
-      expect(new BaseComponent({text40: true}).extractTypographyValue()).toEqual(Typography.text40);
-      expect(new BaseComponent({text70: true}).extractTypographyValue()).toEqual(Typography.text70);
-    });
-
-    it('should return undefined if not typography modifier was sent', () => {
-      expect(new BaseComponent({}).extractTypographyValue()).toEqual(undefined);
-    });
-
-    it('should return take the last typography modifier prop in case there is more than one', () => {
-      expect(new BaseComponent({
-        text40: true,
-        text70: true
-      }).extractTypographyValue(),).toEqual(Typography.text70);
-      expect(new BaseComponent({
-        text70: true,
-        text40: true
-      }).extractTypographyValue(),).toEqual(Typography.text40);
-      expect(new BaseComponent({
-        text40: true,
-        text70: false
-      }).extractTypographyValue(),).toEqual(Typography.text40);
-    });
-
-    it('should return value of the custom made typography', () => {
-      const customTypography = {fontSize: Typography.text30.fontSize, fontWeight: '400'};
-      Typography.loadTypographies({customTypography});
-      expect(new BaseComponent({customTypography: true}).extractTypographyValue()).toEqual(customTypography);
-      expect(new BaseComponent({
-        text40: true,
-        customTypography: true
-      }).extractTypographyValue(),).toEqual(customTypography);
-      expect(new BaseComponent({
-        customTypography: true,
-        text40: true
-      }).extractTypographyValue(),).toEqual(Typography.text40);
-    });
-  });
-
+  
   describe('updateModifiers', () => {
     it('should update state with new modifiers values if modifiers props have changed', () => {
       const uut = new BaseComponent({});
