@@ -2,7 +2,8 @@
 export {default as UIComponent} from './UIComponent';
 export {default as asBaseComponent, BaseComponentInjectedProps} from './asBaseComponent';
 export {default as forwardRef, ForwardRefInjectedProps} from './forwardRef';
-export {default as withScrollEnabler} from './withScrollEnabler';
+export {default as withScrollEnabler, WithScrollEnablerProps} from './withScrollEnabler';
+export {default as withScrollReached, WithScrollReachedProps} from './withScrollReached';
 export {
   ContainerModifiers,
   MarginModifiers,
